CBQE-6252: Separate valid/invalid repository tests
The valid/invalid repository tests bundle together the cloud and the
filesystem types.

When setting the objstore_provider variable, the EnterpriseBaseTestCase
will change archive location to a name that makes more sense for backing
up to the cloud.

Consequently, we can separate the filesystem and cloud cases into
different tests to avoid running into that situation.

Change-Id: I8936cedfd1bf6330daa656f37d6fbe5b50c20547
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141697
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>