CBQE-6251: Amend get_system_time() method to acknowledge the system timezone
Previously, the unix time is fetched and a timezone offset of UTC is
applied to the time. This is coupled with a time reset function which
resets the machine time to UTC at the beginning of every test.

However, couchbase does not acknowledge timezone changes until its
processes are restarted.

Instead of resetting the couchbase server, I could simply change the
get_system_time method to obtain the system time in the rfc3339 format
which contains the timezone information.

We can also remove the timezone reset.

Change-Id: I8b2792950c7866bd0b4d5995d3f63b52fa155ecf
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141798
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>