CBQE-6249: new_install.py marking installation on all VMs as failed incorrectly
This change adds an extra parameter to the exit function to indicate that failures are not due to the install process.
Calls to this function because of a download failure use this new parameter.
This means that only machines that install was attempted on will be marked as failed during install.

Change-Id: I8cf94f84753635a4477efb1beed7d33ec0e21d0e
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141874
Tested-by: Jake Rawsthorne <jake.rawsthorne@couchbase.com>
Reviewed-by: Jagadesh Munta <jagadesh.munta@couchbase.com>