CBQE-6268: Upload root certificate to all nodes.
Previously, the code that was provisioning the custom certificates was
only uploading the root certificate to the first node in the cluster
which resulted in the cluster throwing an error where subsequent nodes
complain that their node level certificate had not been signed by the
root certificate. The fix is to upload the root certificate to all nodes
in the cluster individually.

Change-Id: I38b8e2cffe71e3034b7426a5b6ae54b9a9ce6744
Reviewed-on: http://review.couchbase.org/c/testrunner/+/142691
Reviewed-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>