CBQE-6270: Test with client certificates and enable ssl verification
Supplied ssl hints to the HttpsConfigurationFactory which can configures
the name of the cluster-wide root certificate, client private key and
certificate.

These ssl hints are passed to the HttpsConfigurationFactory to enable
ssl verification.

The class ClientStrategy defines how the username is obtained from the
certificate by Couchbase.

The class ClientCertificateTester assists in generating client private
keys and certificates which have been signed by the cluster's private
key.

Added code to the rest_client to configure and enable client
certificates.

Refactored previous custom certificate into a 'load_custom_certificates'
method. Uses the ClientStrategy and ClientCertificateTester to generate
a key pair for the client.

The custom certificates and custom client certificates can be enabled by
setting the 'custom_certificates' param. Node, the 'use_https' flag must
be supplied.

This can all be coupled with node to node encryption.

Change-Id: If97870af7a1e2a41abef2a9db94eff74d4183438
Reviewed-on: http://review.couchbase.org/c/testrunner/+/142694
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>