CBQE-6273: Amend get_leader() method to receive a list of clusters
The get_leader() method searches the logs for who the current leader is.
The logs of all the nodes in the cluster are searched. If a node is
removed, but present in the list of clusters, the get_leader() method
will return an incorrect leader.

To amend this, we can supply a list of nodes which are currently in the
cluster to get_leadr().

Change-Id: I58d22d3e037f0e327479a22a3922438ddc90700a
Reviewed-on: http://review.couchbase.org/c/testrunner/+/142732
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>