CBQE-6296: Remove execute once in preamble
Remove the execute once clause in the preamble.

The execute once prevents the preamble from run more than once between
multiple test runs which was not original intention.

The original purpose to prevent the preamble from being run twice in a
single test run due to the tearDown being run before the setUp and the
preamble being called in both of those functions.

Change-Id: Ia0530b4c632891f58f884903db0de8c1056ab6e6
Reviewed-on: http://review.couchbase.org/c/testrunner/+/143188
Tested-by: Thuan Nguyen <soccon@gmail.com>
Reviewed-by: Thuan Nguyen <soccon@gmail.com>