CBQE-6327: Use polling to determine when backup service is ready
Using polling to determine when the backup service is ready to respond
to http requests as opposed to sleeping in the killing the backup
service test.

Move the test near to prevent it from interfering with the other tests
in the conf file as it changes the time potentially breaking stats.

Change-Id: I047fa5a4ef88d2cb9743596a5d7d0aceac46ef48
Reviewed-on: http://review.couchbase.org/c/testrunner/+/143402
Reviewed-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>