CBQE-6400: Do not sort the servers.
Sorting the cluster list has unintentional side effects.

Consequently, It's best not to shuffle the elements in these lists of servers.

Sorting was previously done to avoid the sitatuion in which a server of
low memory was being added to a server with high memory causing the add
node to fail.

This is because services share the memory requirements between nodes.
The server with less memory cannot satisfy the memory requirements of
the server with more memory due to the way testrunner configured the
quotas of the services.

Instead, simply assign the correct quotas based on the node with the
lowest available memory before the rebalance happens.

Change-Id: Iffeab22e8f6b3adc4a3ada153039bdeb0308a535
Reviewed-on: http://review.couchbase.org/c/testrunner/+/144647
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>