CBQE-6410: Fix the Samba share
Set the samba_export_all_rw SELinux boolean to prevent SELinux from
preventing the sharing of the folder.

Mount the folder with the couchbase user and group so the backup service
can read and write from it.

Pass the nobrl option in the mount command to prevent the sending of
byte range lock requests to the server.

Change-Id: I975a3d23a50c2764865c668253f2e6a6632eeab6
Reviewed-on: http://review.couchbase.org/c/testrunner/+/144846
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>