CBQE-6889: Re-running jobs with fresh_run=False not working
Check if the same test is already running before calling getservers otherwise should_dispatch_job will always return false. There was previously a bug which caused this to always return true which was accidentally fixed in http://review.couchbase.org/c/testrunner/+/152911. That change also introduced a bug when sdk3 was used because collection.get() returns a different structure to bucket.get(). This change fixes that as well

Change-Id: I7381aab373bfd2503a7e38edc7cbd6188484fb64
Reviewed-on: http://review.couchbase.org/c/testrunner/+/153238
Tested-by: Jake Rawsthorne <jake.rawsthorne@couchbase.com>
Tested-by: Balakumaran G <balakumaran.gopal@couchbase.com>
Reviewed-by: Balakumaran G <balakumaran.gopal@couchbase.com>