CBQE-7116: Installation failing on upgrade tests due to recent changes on install scripts to install debug info
Due to changes in install scripts to install
debug info builds, upgrade tests are
failing on builds where debug info builds aren't
present. Fixing this by not failing the install
if no debug info builds are present. Instead
we will now just print that no debug info builds
are present and we will just move on with the
installation.

Change-Id: Iacfcc98c9881966f68e8f44112f0cce61b159e19
Reviewed-on: http://review.couchbase.org/c/testrunner/+/157435
Reviewed-by: Jagadesh Munta <jagadesh.munta@couchbase.com>
Tested-by: Bharath G P <bharath.gp@couchbase.com>