Optimise base setup and teardown
Optimised base setup and teardown to reduce
redundant code and run the tests faster.
1. Removed multiple cluster resets in tests.
Now the tests will reset the cluster only once,
at the teardown step. Only for first test in
the suite, the setup will be done twice due to
the teardown being called in setup of the test.
2. Removed redundant old legacy code i.e
memcached buckets, sasl buckets and TAP
protocol checks.
3. Removed some dead and obsolete code.

Change-Id: Ib52d506a17fc5e11b9d76026f361de4c39d69815
Reviewed-on: https://review.couchbase.org/c/testrunner/+/165538
Reviewed-by: Jake Rawsthorne <jake.rawsthorne@couchbase.com>
Tested-by: Bharath G P <bharath.gp@couchbase.com>