Fix test failures in eventing due to cluster not being reset after each test
Due to a recent change in testrunner basetest
setup and teardown, we don't run reset cluster
in the setup since it was being run in the teardown
already. But this caused eventing tests that used
the reset cluster in the basetest setup. Moved the
reset cluster to setup as before and removed it from
teardown instead since we need to reset cluster only
once per test.
Also optimised some of the sleeps in reset cluster
function to reduce run times further.

Change-Id: I28469bc6779bbb4dbaad2b5c8e87eacffc24041a
Reviewed-on: https://review.couchbase.org/c/testrunner/+/168041
Tested-by: Couchbase QE
Reviewed-by: Sujay Gad <sujay.gad@couchbase.com>