Fixed merge failures while merging reports
Due to recent changes in basetestcase, a new
test param was added that changes with each run.
This needed to be added to filter test variables
to correctly merge the test reports.

This changes refactored the way we filter test
variables. Now we store the filter expressions
in testconstants so that in the future, we need
to just add new filter expressions in 1 place instead
of adding in multiple places.

Change-Id: I3a5ced4fc32c1a42a86e19fa0b524f17d8fd9b3d
Reviewed-on: https://review.couchbase.org/c/testrunner/+/174022
Reviewed-by: Hemant Kumar Rajput <Hemant.rajput@couchbase.com>
Tested-by: Bharath G P <bharath.gp@couchbase.com>