MB-52400: Add serverless testing for jenkins commit validation
Add support to run simple-test in serverless mode as a starting point
for verification. Obviously there will probably have to be custom
tests created for different uses of serverless, but in the short term
just running the same tests in serverless mode has value, on it's own.
This patch passes the environment from the surrounding build into the
spawned processes that run the server.

Change-Id: I0372bf54917cd603dd0ba9e93ee845c42f5a8ba1
Reviewed-on: https://review.couchbase.org/c/testrunner/+/176129
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by:  Pavithra Mahamani <pavithra.mahamani@couchbase.com>
Reviewed-by: Girish Benakappa <girish.benakappa@couchbase.com>