Fixed installation issues
Due to a recent change in on_prem_rest_client.py,
the initialisation of RestConnection object was
failing when a non test script was trying to
initialise the object. Since in our installation
scripts also uses RestConnection object,
TestInputSingleton object is None and hence the
scripts fail. Fixing this so that install scripts
work.

Change-Id: I67555d715377c6b98a4386389d7df4cae5f19ebd
Reviewed-on: https://review.couchbase.org/c/testrunner/+/188361
Reviewed-by: <ankush.sharma@couchbase.com>
Tested-by: <ankush.sharma@couchbase.com>