CBQE-7879: Dispatcher: Rerunning failed job is rerunning the entire suite instead of just failed ones
Fixed the rerun scripts to rerun in case of
cloud runs too. These runs have parameters
without any values, which is causing the
reruns to fail.
Also added logic to add bucket storage and
GSI types to the tests to differentiate same
jobs but with differing bucket storage or
gsi types.

Change-Id: I692fe21699ab11ed055f847f5f8d5e540869520a
Reviewed-on: https://review.couchbase.org/c/testrunner/+/190022
Reviewed-by: Ashwin <ashwin.govindarajulu@couchbase.com>
Tested-by: Bharath G P <bharath.gp@couchbase.com>